extract crossref tests from itemcmd tests
